- further improvements on showMatrix and showAtomicData - moved legacy code to utilities
